Web/backend/requirements.txt
aerich==0.4.3
aiofiles==0.6.0
aiomysql==0.0.21
aiosqlite==0.13.0
asyncpg==0.21.0
bcrypt==3.2.0
cffi==1.14.6
click==7.1.2
ddlparse==1.9.0
Deprecated==1.2.13
dnspython==2.0.0
email-validator==1.1.1
fastapi==0.68.1
fastapi-crudrouter==0.8.1
fastapi-users==6.1.3
h11==0.9.0
httptools==0.1.1
idna==2.10
iso8601==0.1.12
makefun==1.9.4
passlib==1.7.4
pycparser==2.20
pydantic==1.8.2
PyJWT==2.1.0
PyMySQL==0.9.3
pyparsing==2.4.7
PyPika==0.37.7
python-multipart==0.0.5
six==1.15.0
starlette==0.14.2
tortoise-orm==0.16.13
typing-extensions==3.10.0.2
uvicorn==0.11.5
uvloop==0.14.0
websockets==8.1
wrapt==1.12.1
XlsxWriter==1.3.7
python-dateutil==2.8.2
